refactor: migrate to workspace structure with mbr-core and mbr-cli
- Convert to Cargo workspace with virtual manifest
- Extract shared code to mbr-core crate
- Keep CLI-specific code in mbr-cli crate
- Add prelude module and improve documentation
