+++ Cargo.toml
@@ -1,20 +1,19 @@
-[package]
-name = "mbr-cli"
-version = "0.1.3"
+[workspace]
+members = ["crates/*"]
+resolver = "2"
+
+[workspace.package]
 edition = "2024"
 license = "MIT"
-description = "Rust-based CLI tool for Metabase API interaction"
 repository = "https://github.com/daisuke8000/mbr-cli"
-keywords = ["metabase", "cli", "api", "database", "analytics"]
-categories = ["command-line-utilities", "database"]
 
-[dependencies]
-clap = { version = "4.5", features = ["derive", "env"] }
+[workspace.dependencies]
 tokio = { version = "1.40", features = ["full"] }
 reqwest = { version = "0.11", features = ["json"] }
 serde = { version = "1.0", features = ["derive"] }
 serde_json = "1.0"
 thiserror = "1.0"
+clap = { version = "4.5", features = ["derive", "env"] }
 comfy-table = "7.1"
 toml = "0.8"
 dirs = "6.0"
@@ -26,6 +25,5 @@ crossterm = "0.29"
 atty = "0.2"
 async-trait = "0.1"
 log = "0.4"
-
-[dev-dependencies]
 tempfile = "3.20"
+ratatui = "0.29"
