+++ crates/mbr-core/Cargo.toml
@@ -1,0 +1,27 @@
+[package]
+name = "mbr-core"
+version = "0.1.0"
+edition.workspace = true
+license.workspace = true
+description = "Core library for Metabase API interaction"
+
+[dependencies]
+tokio.workspace = true
+reqwest.workspace = true
+serde.workspace = true
+serde_json.workspace = true
+thiserror.workspace = true
+toml.workspace = true
+dirs.workspace = true
+keyring.workspace = true
+unicode-width.workspace = true
+backoff.workspace = true
+async-trait.workspace = true
+log.workspace = true
+crossterm.workspace = true
+comfy-table.workspace = true
+atty.workspace = true
+rpassword.workspace = true
+
+[dev-dependencies]
+tempfile.workspace = true
