+++ crates/mbr-core/src/display/advanced_pagination.rs
