+++ crates/mbr-core/src/display/pagination.rs
