+++ crates/mbr-core/src/display/table.rs
