+++ crates/mbr-core/src/error.rs
