+++ crates/mbr-core/src/utils/helpers.rs
