+++ crates/mbr-core/src/utils/input.rs
