+++ crates/mbr-core/src/utils/memory.rs
