+++ crates/mbr-core/src/utils/validation.rs
